app/requirements.txt
nltk
Pillow
Flask
flask_cors
mysql-connector-python